requirements.txt
Django==6.0.4
gunicorn==25.3.0
pillow==12.2.0
whitenoise==6.12.0
